Refactor Contentful libs and foldering
