+++ src/app/api/contentful/pages/[slug]/route.ts
@@ -1,14 +1,14 @@
-import { getPages } from "@/src/lib/route-handler/contentful";
+import { getPageBySlug } from "@/src/lib/contentful/get-pages";
 import {
   type ParamsProps,
   nextResponseSuccess,
   nextResponseFailed,
-} from "@/src/lib/route-handler/utils";
+} from "@/src/lib/route-handler";
 
 export async function GET(_: Request, { params }: ParamsProps) {
   try {
     const { slug } = await params;
-    const pageBySlug = await getPages(slug);
+    const pageBySlug = await getPageBySlug(slug);
     return nextResponseSuccess(pageBySlug);
   } catch (err) {
     return nextResponseFailed(err, "contentful/pages/[slug]");
