+++ src/app/api/contentful/pages/route.ts
@@ -1,9 +1,9 @@
-import { getPages } from "@/src/lib/route-handler/contentful";
-import { mapAllSlugs } from "@/src/lib/contentful";
+import { getPages } from "@/src/lib/contentful/get-pages";
+import { mapAllSlugs } from "@/src/lib/contentful/utils";
 import {
   nextResponseSuccess,
   nextResponseFailed,
-} from "@/src/lib/route-handler/utils";
+} from "@/src/lib/route-handler";
 
 export async function GET() {
   try {
