+++ src/constants/contentful.ts
@@ -1,7 +1,11 @@
+// Common setup
+export const DEFAULT_LOCALE = "en-US";
+export const ENTRIES_LEVELS = 3;
+
+// Content type
 export const CONTENT_TYPE_PAGE = "page";
 export const CONTENT_TYPE_HERO = "heroBanner";
 export const CONTENT_TYPE_FEATURE = "featureBanner";
 
-// Contentful SDK setup
-export const DEFAULT_LOCALE = "en-US";
-export const ENTRIES_LEVELS = 3;
+// Fields
+export const FIELD_PAGE_SLUG = "fields.slug";
