+++ src/lib/contentful/utils.ts
@@ -1,4 +1,3 @@
-import { getPages } from "@/src/lib/route-handler/contentful";
 import { ContentfulRoot } from "@/src/type/contentful";
 
 const stripLeadingSlash = (slug: string): string =>
@@ -6,10 +5,3 @@ const stripLeadingSlash = (slug: string): string =>
 
 export const mapAllSlugs = (allPages: ContentfulRoot) =>
   allPages.items.map((page) => stripLeadingSlash(page.fields.slug));
-
-export const getPageSlugs = async () => {
-  const allPages = await getPages();
-  return mapAllSlugs(allPages);
-};
-
-export const getPageBySlug = async (slug: string) => getPages(slug);
